\name{adoptees}
\docType{data}
\alias{adoptees}
\title{Adoptees Education}
\description{a cross-section of 16481 individuals from 1992

    \emph{country} :  United States 

    \emph{package} : durationbinomial

    \emph{JEL codes: }I21, J13, J16

}
\usage{data(adoptees)}
\format{A dataframe containing:
  \describe{
    \item{adopt}{is the child adopted ?}
    \item{educ}{years of education}
    \item{coll}{college graduate}
    \item{stinsch}{still in school}
    \item{gender}{one of 'male' and 'female'}
    \item{age}{the age in years}
    \item{feduc}{father's education}
    \item{meduc}{mother's education}
    \item{fcoll}{is the father college graduate ?}
    \item{mcoll}{is the mother college graduate ?}
    \item{income}{income}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Plug, Erik (2004) \dQuote{Estimating the Effect of Mother's Schooling on Children's Schooling Using a Sample of Adoptees}, \emph{American Economic Review}, \bold{94(1)}, 358-368.

}
\keyword{datasets}
