\name{afghan_girls}
\docType{data}
\alias{afghan_girls}
\title{Bringing Education to Afghan Girls}
\description{a cross-section of 1490 individuals from 2007

    \emph{country} :  Afghanistan 

    \emph{package} : 

}
\usage{data(afghan_girls)}
\format{A dataframe containing:
  \describe{
    \item{group}{one of `"control"` and `"treatment"`}
    \item{village}{village groups' id}
    \item{enrollment}{enrollment in formal school, one of `"no"` and `"yes"`}
    \item{test}{standardized test}
    \item{chagcharan}{unknown covariate}
    \item{head_child}{head's child, one of `"no"` and `"yes"`}
    \item{sex}{sex, one of `"boy"` and `"girl"`}
    \item{age}{age of the child}
    \item{duration}{length of stay of the family in the village}
    \item{occup}{occupation of the head of the household, one of `"other"` and `"farmer"`}
    \item{age_head}{age of the head of the household}
    \item{educ}{education level of the head of the household}
    \item{hsize}{household's size}
    \item{jeribs}{jeribs of land}
    \item{sheeps}{number of sheep}
    \item{distance}{distance to nearest formal scholl}
    \item{ethny}{etnicity one of `"other"`, `"farsi"` and `"tajik"`}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Burde, Dana and Leigh L. Linden (2013) \dQuote{Bringing Education to Afghan Girls: a Randomized Controlled Trial of Village-based Schools}, \emph{American Economic Journal: Applied Economics}, \bold{5(3)}, 27-40.

}
\keyword{datasets}
