\name{agglo_growth}
\docType{data}
\alias{agglo_growth}
\title{Economies and Diseconomies of Scale for Us Counties}
\description{a cross-section of 3106 counties from 1990

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(agglo_growth)}
\format{A dataframe containing:
  \describe{
    \item{fips}{county fips code}
    \item{emp_gr}{employment growth rate}
    \item{pop_gr}{population growth rate}
    \item{emp}{employment in 1980}
    \item{pop}{population in 1980}
    \item{college}{fraction of adult population with bachelor's degree or more in 1980}
    \item{manuf}{fraction of employment in manufacturing in 1980}
    \item{unemp}{unemployment rate in 1980}
    \item{income}{per capita income in 1979}
    \item{educ_sh}{share of local government spending on education in 1982}
    \item{hw_sh}{share of local government spending on highways in 1982}
    \item{pol_sh}{share of local government spending on police in 1982}
    \item{notwhite}{fraction of population that is not white in 1980}
    \item{type}{a factor with levels `"rural"` and `"urban"`}
  }
}
\source{JAE data archive}
\references{
Wheeler, Christopher H. (2003) \dQuote{Evidence on Agglomeration Economies, Diseconomies, and Growth}, \emph{Journal of Applied Econometrics}, \bold{18(1)}, 79-104.

}
\keyword{datasets}
