\name{apples}
\docType{data}
\alias{apples}
\title{Apple Production}
\description{yearly observations of 173 farms from 1984 to 1986

    \emph{number of observations} : 405

    \emph{number of  time-series } :  1-3 

    \emph{country} :  France 

    \emph{package} : productionpanel

    \emph{JEL codes: }C51, D24, Q12

}
\usage{data(apples)}
\format{A dataframe containing:
  \describe{
    \item{id}{farm's id}
    \item{year}{year}
    \item{capital}{capital stock}
    \item{labor}{quantity of labor}
    \item{materials}{quantity of materials}
    \item{apples}{production of apples}
    \item{otherprod}{other productions}
    \item{pc}{price of capital}
    \item{pl}{price of labor}
    \item{pm}{price of materials}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Ivaldi, Marc; Ladoux, Norbert; Ossard, Herv\'e and Michel Simioni (1996) \dQuote{Comparing Fourier and Translog Specifications of Multiproduct Technology: Evidence From An Incomplete Panel of French Farmers}, \emph{Journal of Applied Econometrics}, \bold{11(6)}, 649--668.

}
\keyword{datasets}
