\name{basque_country}
\docType{data}
\alias{basque_country}
\title{Economic Effects of Terrorism in the Basque Country}
\description{a pseudo-panel of 17 regions from 1955 to 1997

    \emph{number of observations} : 731

    \emph{number of  individual observations } :  43 

    \emph{country} :  Spain 

    \emph{package} : 

}
\usage{data(basque_country)}
\format{A dataframe containing:
  \describe{
    \item{id}{region's id}
    \item{region}{regions's name}
    \item{year}{the year}
    \item{gdpcap}{gdp per capita}
    \item{agriculture}{NA}
    \item{energy}{}
    \item{industry}{NA}
    \item{construction}{NA}
    \item{services}{NA}
    \item{administration}{NA}
    \item{illit_educ}{NA}
    \item{prim_educ}{NA}
    \item{medium_educ}{NA}
    \item{high_educ}{NA}
    \item{popdens}{population density}
  }
}
\source{Synth package}
\references{
Abadie, Alberto and Javier Gardeazabal (2003) \dQuote{The Economic Costs of Conflict: A Case Study of the Basque Country}, \emph{American Economic Review}, \bold{93(1)}, 113-132.

}
\keyword{datasets}
