\name{bdh}
\docType{data}
\alias{bdh}
\title{Income and Religiousness: Bono Desarollo Humano Program}
\description{a cross-section of 2645 individuals from 2011

    \emph{country} :  Equador 

    \emph{package} : 

}
\usage{data(bdh)}
\format{A dataframe containing:
  \describe{
    \item{ciudad}{city name, one of `"guayaquil"`, `"quito"` and `"sto domingo"`}
    \item{parroqui}{parish name (one out of 36)}
    \item{selben}{centered score on the selben index (the threshold is 0)}
    \item{treated}{1 if treated}
    \item{previous_elig}{1 if previously eligible}
    \item{religion}{one of `"other"`, `"catholic"`, `"other_christian"`, `"jewish"` and `"atheist"`}
    \item{attendance}{attendance at church, a factor}
    \item{attendmonth}{monthly visits to church}
    \item{religiousness}{self-rated religiousness, from 0 to 10}
    \item{age}{respondent's age}
    \item{hsize}{household's size}
    \item{schooling}{respondent's number of years of schooling}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Buser, Thomas (2015) \dQuote{The Effect of Income One Religiousness}, \emph{American Economic Journal: Applied Economics}, \bold{7(3)}, 178-195.

}
\keyword{datasets}
