\name{birthwt}
\docType{data}
\alias{birthwt}
\title{Cigarette Smoking and Birth Weight}
\description{a cross-section of 1388 individuals from 1988

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(birthwt)}
\format{A dataframe containing:
  \describe{
    \item{birthwt}{birth weight}
    \item{cigarettes}{number of cigarettes smoked per day during pregnancy}
    \item{parity}{birth order}
    \item{race}{a factor with levels other and white}
    \item{sex}{a factor with levels female and male}
    \item{edmother}{number of years of education of the mother}
    \item{edfather}{number of years of education of the father}
    \item{faminc}{family income}
    \item{cigtax}{per-pack state excise tax on cigarettes}
  }
}
\source{kindly provided by John Mullahy}
\references{
Mullahy, John (1997) \dQuote{Instrumental-variable Estimation of Count Data Models: Applications to Models of Cigarette Smoking Behavior}, \emph{Review of Economics and Statistics}, \bold{79(4)}, 586-593.

}
\keyword{datasets}
