\name{food}
\docType{data}
\alias{food}
\title{Food Consumption in the Netherlands}
\description{a cross-section of 4611 households

    \emph{country} :  Netherlands 

    \emph{package} : 

    \emph{JEL codes: }D12, C50

}
\usage{data(food)}
\format{A dataframe containing:
  \describe{
    \item{year}{year of the survey, either 1980 or 1988}
    \item{weights}{weight in the survey}
    \item{hsize}{number of persons in the household}
    \item{ageh}{age of the head of the household in 5 classes}
    \item{income}{total income}
    \item{food}{food expenditure}
    \item{midage}{a dummy for household for which the head is aged between 35 and 64 years old}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
CromBrugghe, Denis de; Palm, Franz C. and Jean-Pierre Urbain (1997) \dQuote{Statistical Demand Functions for Food in the USA and the Netherlands}, \emph{Journal of Applied Econometrics}, \bold{12(5)}, 615-645, \doi{10.1002/(SICI)1099-1255(199709/10)12:5<615::AID-JAE455>3.0.CO;2-L}
.

}
\keyword{datasets}
