\name{Index.economics}
\docType{data}
\alias{Index.Economic Topics}
\title{Economic Topics}
\description{
\itemize{
\item   
\itemize{
\item \code{\link{basque_country}} : Economic effects of terrorism in the Basque country
\item \code{\link{growth}} : Augmented Sollow's growth model with human capital
\item \code{\link{paces}} : Voucher and private school
\item \code{\link{portfolio}} : Precautionary motives and portfolio decisions
\item \code{\link{sibling_educ}} : Education of oldest sibling as an instrument of own education
\item \code{\link{telephone}} : Choice among residential telephone service options for local calling
\item \code{\link{unemp_duration}} : Unemployment Duration in Germany
\item \code{\link{unemp_teachers}} : Unemployment of teachers in the Netherlands
\item \code{\link{wine}} : Price of wine
}
\item bank  
\itemize{
\item \code{\link{mortgage_defaults}} : Mortgage Defaults
}
\item consumer  
\itemize{
\item \code{\link{napster}} : Napster and expense for recorded music
}
\item consumer behaviour  
\itemize{
\item \code{\link{dutch_railways}} : Stated preferences for train traveling
\item \code{\link{uk_elec}} : Electricity consumption in the United Kingdom
}
\item contingent valuation  
\itemize{
\item \code{\link{kakadu}} : Willingness to pay for the preservation of the Kakadu national park
}
\item crime economics  
\itemize{
\item \code{\link{recidivism}} : recidivism
}
\item development economics  
\itemize{
\item \code{\link{slave_trade}} : Slave trade and economic development
}
\item economics of crime  
\itemize{
\item \code{\link{car_thefts}} : Police presence and thefts
}
\item education  
\itemize{
\item \code{\link{probation}} : Academic probation
}
\item growth  
\itemize{
\item \code{\link{sp_solow}} : Solow's growth model with spatial correlation
}
\item health economics  
\itemize{
\item \code{\link{birthwt}} : Cigarette smoking and birth weight
}
\item income distribution  
\itemize{
\item \code{\link{kuznets}} : Kuznets curve
}
\item industrial economics  
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item labor economics  
\itemize{
\item \code{\link{adoptees}} : Adoptees education 
\item \code{\link{disregard}} : Unemployment insurance rules
\item \code{\link{strikes}} : strikes frequency and duration
}
\item labor Economics   
\itemize{
\item \code{\link{twins}} : Education and earnings of british twins
}
\item micro-economics  
\itemize{
\item \code{\link{airbnb}} : Racial Discrimination in the Sharing Economy
\item \code{\link{food}} : Food consumption in the Netherlands
\item \code{\link{loan_market}} : US commercial loan market
}
\item microeconomics  
\itemize{
\item \code{\link{retirement}} : Joint retirement of married couples
\item \code{\link{tobinq}} : The Q theory of investment
}
\item microeconomics   
\itemize{
\item \code{\link{watermelon}} : Equilibrium on the watermelon market
}
\item none  
\itemize{
\item \code{\link{us_counties}} : US Counties
}
\item Political Economy  
\itemize{
\item \code{\link{traffic_citations}} : Political Economy and Traffic Citations
}
\item producer behavior  
\itemize{
\item \code{\link{apples}} : Apple production
\item \code{\link{oil}} : Oil investment
}
\item social  
\itemize{
\item \code{\link{bdh}} : Income and Religiousness: Bono Desarollo Humano program
}
\item transport economics  
\itemize{
\item \code{\link{price_time}} : Market shares of train and air for Paris and main French cities in 1995
\item \code{\link{toronto_montreal}} : Mode Choice for the Toronto-Montreal Corridor
}
\item treatment effect  
\itemize{
\item \code{\link{afghan_girls}} : Bringing education to Afghan girls
}
\item unknown  
\itemize{
\item \code{\link{vietnam}} : Draft eligibility and income
}
\item urban economics  
\itemize{
\item \code{\link{agglo_growth}} : Economies and diseconomies of scale for US counties
\item \code{\link{tracks_side}} : Railroad tracks and segregation
\item \code{\link{urban_gradient}} : Density and distance to the CBD in Alabama
}
}
}
\keyword{documentation}
