\name{Index.jel}
\docType{data}
\alias{Index.JEL codes}
\title{JEL codes}
\description{
\itemize{
\item \bold{ 2120 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ 6310 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ 6352 } :  NA 
\itemize{
\item \code{\link{telephone}} : Choice among residential telephone service options for local calling
}
\item \bold{ 8242 } :  NA 
\itemize{
\item \code{\link{vietnam}} : Draft eligibility and income
}
\item \bold{ 8321 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ 9212 } :  NA 
\itemize{
\item \code{\link{telephone}} : Choice among residential telephone service options for local calling
}
\item \bold{ C14 } :  Semiparametric and Nonparametric Methods: General 
\itemize{
\item \code{\link{sibling_educ}} : Education of oldest sibling as an instrument of own education
}
\item \bold{ C25 } :  Single Equation Models; Single Variables: Discrete Regression and Qualitative Choice Models; Discrete Regressors; Proportions; Probabilities 
\itemize{
\item \code{\link{kakadu}} : Willingness to pay for the preservation of the Kakadu national park
\item \code{\link{toronto_montreal}} : Mode Choice for the Toronto-Montreal Corridor
}
\item \bold{ C30 } :  Multiple or Simultaneous Equation Models; Multiple Variables: General 
\itemize{
\item \code{\link{sibling_educ}} : Education of oldest sibling as an instrument of own education
}
\item \bold{ C50 } :  Econometric Modeling: General 
\itemize{
\item \code{\link{food}} : Food consumption in the Netherlands
}
\item \bold{ C51 } :  Model Construction and Estimation 
\itemize{
\item \code{\link{apples}} : Apple production
\item \code{\link{dutch_railways}} : Stated preferences for train traveling
\item \code{\link{kakadu}} : Willingness to pay for the preservation of the Kakadu national park
\item \code{\link{oil}} : Oil investment
\item \code{\link{wine}} : Price of wine
}
\item \bold{ C93 } :  Field Experiments 
\itemize{
\item \code{\link{airbnb}} : Racial Discrimination in the Sharing Economy
}
\item \bold{ D12 } :  Consumer Economics: Empirical Analysis 
\itemize{
\item \code{\link{dutch_railways}} : Stated preferences for train traveling
\item \code{\link{food}} : Food consumption in the Netherlands
}
\item \bold{ D24 } :  Production; Cost; Capital; Capital, Total Factor, and Multifactor Productivity; Capacity 
\itemize{
\item \code{\link{apples}} : Apple production
}
\item \bold{ D43 } :  Market Structure, Pricing, and Design: Oligopoly and Other Forms of Market Imperfection 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ G11 } :  Portfolio Choice; Investment Decisions 
\itemize{
\item \code{\link{portfolio}} : Precautionary motives and portfolio decisions
}
\item \bold{ G31 } :  Capital Budgeting; Fixed Investment and Inventory Studies; Capacity 
\itemize{
\item \code{\link{oil}} : Oil investment
}
\item \bold{ H76 } :  State and Local Government: Other Expenditure Categories 
\itemize{
\item \code{\link{traffic_citations}} : Political Economy and Traffic Citations
}
\item \bold{ I21 } :  Analysis of Education 
\itemize{
\item \code{\link{adoptees}} : Adoptees education 
\item \code{\link{twins}} : Education and earnings of british twins
}
\item \bold{ I32 } :  Measurement and Analysis of Poverty 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ J13 } :  Fertility; Family Planning; Child Care; Children; Youth 
\itemize{
\item \code{\link{adoptees}} : Adoptees education 
}
\item \bold{ J15 } :  Economics of Minorities, Races, Indigenous Peoples, and Immigrants; Non-labor Discrimination 
\itemize{
\item \code{\link{airbnb}} : Racial Discrimination in the Sharing Economy
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ J16 } :  Economics of Gender; Non-labor Discrimination 
\itemize{
\item \code{\link{adoptees}} : Adoptees education 
}
\item \bold{ J24 } :  Human Capital; Skills; Occupational Choice; Labor Productivity 
\itemize{
\item \code{\link{twins}} : Education and earnings of british twins
}
\item \bold{ J31 } :  Wage Level and Structure; Wage Differentials 
\itemize{
\item \code{\link{sibling_educ}} : Education of oldest sibling as an instrument of own education
}
\item \bold{ J45 } :  Public Sector Labor Markets 
\itemize{
\item \code{\link{car_thefts}} : Police presence and thefts
}
\item \bold{ K21 } :  Antitrust Law 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ K42 } :  Illegal Behavior and the Enforcement of Law 
\itemize{
\item \code{\link{car_thefts}} : Police presence and thefts
}
\item \bold{ L11 } :  Production, Pricing, and Market Structure; Size Distribution of Firms 
\itemize{
\item \code{\link{wine}} : Price of wine
}
\item \bold{ L12 } :  Monopoly; Monopolization Strategies 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ L13 } :  Oligopoly and Other Imperfect Markets 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ L66 } :  Food; Beverages; Cosmetics; Tobacco; Wine and Spirits 
\itemize{
\item \code{\link{wine}} : Price of wine
}
\item \bold{ L71 } :  Mining, Extraction, and Refining: Hydrocarbon Fuels 
\itemize{
\item \code{\link{oil}} : Oil investment
}
\item \bold{ L83 } :  Sports; Gambling; Restaurants; Recreation; Tourism 
\itemize{
\item \code{\link{airbnb}} : Racial Discrimination in the Sharing Economy
}
\item \bold{ N31 } :  Economic History: Labor and Consumers, Demography, Education, Health, Welfare, Income, Wealth, Religion, and Philanthropy: U.S.; Canada: Pre-1913 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ N32 } :  Economic History: Labor and Consumers, Demography, Education, Health, Welfare, Income, Wealth, Religion, and Philanthropy: U.S.; Canada: 1913- 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ N37 } :  Economic History: Labor and Consumers, Demography, Education, Health, Welfare, Income, Wealth, Religion, and Philanthropy: Africa; Oceania 
\itemize{
\item \code{\link{slave_trade}} : Slave trade and economic development
}
\item \bold{ N91 } :  Regional and Urban History: U.S.; Canada: Pre-1913 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ N92 } :  Regional and Urban History: U.S.; Canada: 1913- 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ O10 } :  Economic Development: General 
\itemize{
\item \code{\link{slave_trade}} : Slave trade and economic development
}
\item \bold{ O17 } :  Formal and Informal Sectors; Shadow Economy; Institutional Arrangements 
\itemize{
\item \code{\link{car_thefts}} : Police presence and thefts
}
\item \bold{ O41 } :  One, Two, and Multisector Growth Models 
\itemize{
\item \code{\link{growth}} : Augmented Sollow's growth model with human capital
}
\item \bold{ O47 } :  Empirical Studies of Economic Growth; Aggregate Productivity; Cross-Country Output Convergence 
\itemize{
\item \code{\link{growth}} : Augmented Sollow's growth model with human capital
}
\item \bold{ Q12 } :  Micro Analysis of Farm Firms, Farm Households, and Farm Input Markets 
\itemize{
\item \code{\link{apples}} : Apple production
}
\item \bold{ Q26 } :  Recreational Aspects of Natural Resources 
\itemize{
\item \code{\link{kakadu}} : Willingness to pay for the preservation of the Kakadu national park
}
\item \bold{ R23 } :  Urban, Rural, Regional, Real Estate, and Transportation Economics: Regional Migration; Regional Labor Markets; Population; Neighborhood Characteristics 
\itemize{
\item \code{\link{tracks_side}} : Railroad tracks and segregation
}
\item \bold{ R41 } :  Transportation: Demand, Supply, and Congestion; Travel Time; Safety and Accidents; Transportation Noise 
\itemize{
\item \code{\link{toronto_montreal}} : Mode Choice for the Toronto-Montreal Corridor
\item \code{\link{traffic_citations}} : Political Economy and Traffic Citations
}
}
}
\keyword{documentation}
