\name{kuznets}
\docType{data}
\alias{kuznets}
\title{Kuznets Curve}
\description{a cross-section of 60 countries

    \emph{country} :  world 

    \emph{package} : 

}
\usage{data(kuznets)}
\format{A dataframe containing:
  \describe{
    \item{country}{the country}
    \item{group}{group of countries, one of `"developing"`, `"developed"` and `"socialist"`}
    \item{gnp}{gross national product in 1979 US dollars}
    \item{gini}{gini index}
    \item{varlog}{variance of log income}
  }
}
\source{table A.1 in the original paper}
\references{
Anand, Sudhir and S.M.R. Kanbur (1993) \dQuote{The Kuznets Process and the Inequality-development Relationship}, \emph{Journal of Development Economics}, \bold{40}, 25-52.

}
\keyword{datasets}
