\name{loan_market}
\docType{data}
\alias{loan_market}
\title{Us Commercial Loan Market}
\description{a time series of 72 monthly observations from 1979-01 to 1984-12

    \emph{country} :  United States 

    \emph{package} : time-series

}
\usage{data(loan_market)}
\format{A dataframe containing:
  \describe{
    \item{date}{the month of observation}
    \item{loans}{total commercial loans, billions of dollars}
    \item{prime_rate}{average prime rate charged by banks}
    \item{aaa_rate}{AAA corporate bond rate (alternative financing for firms)}
    \item{ipi}{industrial production index (firms' expectation about future economic activity)}
    \item{treas_rate}{3-month treasury bill rate (alternative rate of return for banks)}
    \item{deposits}{total bank deposit (scale variable), billions of dollars}
  }
}
\source{G.S. Maddala thanks Water Mayer for providing him with the data (p.364-65)}
\references{
Maddala, G.S.(2001) Introduction to Econometrics, John Wiley and sons, ltd..

}
\keyword{datasets}
