\name{mortgage_defaults}
\docType{data}
\alias{mortgage_defaults}
\title{Mortgage Defaults}
\description{yearly observations of 457 counties from 2007 to 2011

    \emph{number of observations} : 2252

    \emph{number of  time-series } :  5 

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(mortgage_defaults)}
\format{A dataframe containing:
  \describe{
    \item{fips}{county id}
    \item{year}{year}
    \item{default}{rate of mortgage default}
    \item{fico}{fico}
    \item{impl1}{impl1}
    \item{ur}{ur}
    \item{loans}{loans}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Kumar Anil (2018) \dQuote{Do Restrictions on Home Equity Extraction Contribute to Lower Mortgage Defaults? Evidence From a Policy Discontinuity at the Texas Border}, \emph{American Economic Journal: Economic Policy}, \bold{10(1)}, 268-297.

}
\keyword{datasets}
