\name{paces}
\docType{data}
\alias{paces}
\title{Voucher and Private School}
\description{a cross-section of 1577 households from 1995

    \emph{country} :  Colombia 

    \emph{package} : 

}
\usage{data(paces)}
\format{A dataframe containing:
  \describe{
    \item{id}{NA}
    \item{privsch}{NA}
    \item{educyrs}{NA}
    \item{voucher}{NA}
    \item{pilot}{NA}
    \item{housvisit}{NA}
    \item{city}{NA}
    \item{phone}{NA}
    \item{age}{NA}
    \item{sex}{NA}
    \item{strata}{NA}
    \item{smpl}{NA}
    \item{month}{NA}
    \item{married}{NA}
    \item{finish8}{V}
    \item{repetitions}{NA}
    \item{in_school}{NA}
    \item{year}{NA}
  }
}
\source{Joshua Angrist's web site}
\references{
Angrist, Joshua; Bettinger, Eric; Bloom, Erik; King, Elizabeth and Michael Kremer (2002) \dQuote{Vouchers for Private Schooling in Colombia: Evidence From a Randomized Natural Experiment}, \emph{American Economic Review}, \bold{95(2)}, 1535-1558.

}
\keyword{datasets}
