\name{price_time}
\docType{data}
\alias{price_time}
\title{Market Shares of Train and Air for Paris and Main French Cities in 1995}
\description{a cross-section of 13 cities from 1995

    \emph{country} :  France 

    \emph{package} : 

}
\usage{data(price_time)}
\format{A dataframe containing:
  \describe{
    \item{town}{the city}
    \item{trafic_rail}{number of rail passengers}
    \item{trafic_air}{number of air passengers}
    \item{price_rail}{price of a train ticket in euros}
    \item{price_air}{price of a plane ticket in euros}
    \item{time_rail}{length of a train trip in minutes}
    \item{time_air}{length a plane trip in minutes}
  }
}
\source{Patrick Bonnel's book}
\references{
Bonnel, Patrick(2004) Pr{'e}voir la demande de transport, Presses de l'Ecole Nationale des Ponts et Chauss{'e}es, Paris.

}
\keyword{datasets}
