\name{probation}
\docType{data}
\alias{probation}
\title{Academic Probation}
\description{a cross-section of 44362 individuals from 2004

    \emph{country} :  Canada 

    \emph{package} : 

}
\usage{data(probation)}
\format{A dataframe containing:
  \describe{
    \item{hsgrade}{high school grade percentile}
    \item{gender}{a factor with levels `"female"` and `"male"`}
    \item{flang}{first language, one of `"other"` and `"english"`}
    \item{age}{age at entry}
    \item{bplace}{place of birth}
    \item{credits1}{credits attempted at first year}
    \item{gpa}{distance to the cutoff}
    \item{probation}{a dummy for probation}
    \item{ever_susp}{one if ever suspended}
    \item{ever_prob}{one if ever probation}
    \item{dropout}{left university after 1st evaluation}
    \item{gradin4}{graduated by year 4}
    \item{gradin5}{graduated by year 4}
    \item{gradin6}{graduated by year 6}
    \item{gpa2}{distance from cutoff at next evaluation}
    \item{campus}{one of `"1"`, `"2"` and `"3"`}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Lindo, Jason M.; Sanders, Nichols J. and Philip Oreopoulos (2010) \dQuote{Ability, Gender, and Performance Standards: Evidence From Academic Probation}, \emph{American Economic Journal: Applied Economics}, \bold{2(2)}, 95-117.

}
\keyword{datasets}
