\name{sibling_educ}
\docType{data}
\alias{sibling_educ}
\title{Education of Oldest Sibling As An Instrument of Own Education}
\description{a cross-section of 303 individuals from 1992

    \emph{country} :  United States 

    \emph{package} : mincer

    \emph{JEL codes: }C14, C30, J31

}
\usage{data(sibling_educ)}
\format{A dataframe containing:
  \describe{
    \item{wage}{hourly wage}
    \item{educ}{Education (years of schooling)}
    \item{ability}{ability based on standardized AFQT test score}
    \item{educsib}{education of oldest sibling (years of schooling)}
    \item{experience}{experience (total weeks of labor market experience)}
    \item{tenure}{tenure (weeks on current job)}
    \item{mothed}{Mother's education (years of schooling)}
    \item{fathed}{Father's education (years of schooling)}
    \item{residence}{one of urban or rural}
    \item{broken}{broken home dummy (dummy for living with both parents at age 14)}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Koop, Gary; Poirier, Dale and Justin L. Tobias (2005) \dQuote{Semiparametric Bayesian Inference in Multiple Equation Models}, \emph{Journal of Applied Econometrics}, \bold{20(6)}, 723-747, \doi{10.1002/jae.810}
.

}
\keyword{datasets}
