\name{sp_solow}
\docType{data}
\alias{sp_solow}
\title{Solow's Growth Model with Spatial Correlation}
\description{a cross-section of 91 countries from 1995

    \emph{country} :  world 

    \emph{package} : 

}
\usage{data(sp_solow)}
\format{A dataframe containing:
  \describe{
    \item{name}{the name of the country}
    \item{code}{the id of the country}
    \item{capital}{name of the capital}
    \item{gdp60}{per capita gdp in 1960}
    \item{gdp95}{per capita gdp in 1995}
    \item{saving}{saving rate}
    \item{labgwth}{growth rate of the labor force}
    \item{point}{the coordinates of the capital}
    \item{border}{the coordinates of the country}
  }
}
\source{JAE data archive}
\references{
Ertur, Cem and Wilfried Koch (2007) \dQuote{Growth, Technological Interdependence and Spatial Externalities: Theory and Evidence}, \emph{Journal of Applied Econometrics}, \bold{22(6)}, 1033--1062.

}
\keyword{datasets}
