\name{strikes}
\docType{data}
\alias{strikes}
\title{Strikes Frequency and Duration}
\description{a cross-section of 573  from 1968-01 to 1976-12

    \emph{country} :  

    \emph{package} : 

    \emph{JEL codes: }2120, 6310, 8321

}
\usage{data(strikes)}
\format{A dataframe containing:
  \describe{
    \item{month}{a date, the first day of the month of observation}
    \item{strikes}{the number of strikes}
    \item{output}{potential output}
    \item{duration}{duration of the strike in days}
  }
}
\source{the original paper, pages 14-16}
\references{
Kennan, John (1985) \dQuote{The Duration of Contract Strikes in Us Manufacturing}, \emph{Journal of Econometrics}, \bold{28}, 5-28.

}
\keyword{datasets}
