\name{tracks_side}
\docType{data}
\alias{tracks_side}
\title{Railroad Tracks and Segregation}
\description{a cross-section of 121 cities from 1990

    \emph{country} :  United States 

    \emph{package} : ivreg

    \emph{JEL codes: }I32, J15, N31, N32, N91, N92, R23

}
\usage{data(tracks_side)}
\format{A dataframe containing:
  \describe{
    \item{city}{city}
    \item{state}{state}
    \item{segregation}{dissimilarity index in 1990 of the city}
    \item{trackdens}{track length per square kilometer}
    \item{raildiv}{railroad division index}
    \item{giniw}{gini index for whites}
    \item{ginib}{gini index for blacks}
    \item{povw}{poverty rate for whites}
    \item{povb}{poverty rate for blacks}
    \item{D9wD9b}{ratio of the ninth decile for whites and blacks}
    \item{D1wD1b}{ratio of the first decile for whites and blacks}
    \item{D9wD1b}{ratio of the ninth decile for whites and the first decile for blacks}
    \item{D9bD1w}{ratio of the ninth decile for blacks and the first decile for whites}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Ananat, Elizabeth Oltmans (2011) \dQuote{The Wrong Side(s) of the Tracks: the Causal Effects of Racial Segregation on Urban Poverty and Inequality}, \emph{American Economic Journal: Applied Economics}, \bold{3(2)}, 34-66, \doi{10.1257/app.3.2.34}
.

}
\keyword{datasets}
