\name{twins}
\docType{data}
\alias{twins}
\title{Education and Earnings of British Twins}
\description{a pseudo-panel of 214 families from 1999

    \emph{number of observations} : 428

    \emph{number of  individual observations } :  2 

    \emph{country} :  United-Kingdown 

    \emph{package} : mincer

    \emph{JEL codes: }I21, J24

}
\usage{data(twins)}
\format{A dataframe containing:
  \describe{
    \item{family}{family index}
    \item{twin}{twin index (1 or 2)}
    \item{earning}{hourly wage rate}
    \item{decl_educ}{reported years of schooling}
    \item{educ}{estimated years of schooling (base on the highest qualification)}
    \item{educt}{estimated years of schooling declared by the twin}
    \item{age}{age}
    \item{londonse}{lives in London or in the South-East}
    \item{part}{part-time job ?}
    \item{self}{self-emploied ?}
    \item{married}{married ?}
    \item{bweight}{birth weight}
    \item{smoking16}{smoking at 16 years old ?}
    \item{smoking18}{smoking at 18 years old ?}
    \item{tenure}{tenure in years}
    \item{occ_part}{occupation of the partner, a factor with 8 levels (plant and machine operatives, sales occupations, personal and protective services, crafts and related occupations, clerical and secretarial occupations, associate professional occupations, professional occupations)}
    \item{tenure_part}{tenure of the partner in years}
  }
}
\source{American Economic Review data archive : \url{http://aeaweb.org/aer/}}
\references{
Bonjour, D.; Cherkas, L.F.; Haskel, J.E.; Hawkes, D.D. and T.D. Spector (2003) \dQuote{Returns to Education: Evidence From U.k. Twins}, \emph{American Economic Review}, \bold{93(5)}, 1799-1812.

}
\keyword{datasets}
