\name{uk_elec}
\docType{data}
\alias{uk_elec}
\title{Electricity Consumption in the United Kingdom}
\description{a cross-section of 42 cities

    \emph{country} :  United Kingdom 

    \emph{package} : 

}
\usage{data(uk_elec)}
\format{A dataframe containing:
  \describe{
    \item{city}{the city}
    \item{cust}{Average number of consumers with two-part tariffs in 1937-38, in thousands}
    \item{inc}{Average income of two-part consumers, in pounds per year}
    \item{mc4}{The running charge (marginal cost) on domestic two-part tariffs in 1933-34, in pence per KWH'}
    \item{mc6}{The running charge (marginal cost) on domestic two-part tariffs in 1935-36, in pence per KWH'}
    \item{mc8}{The running charge (marginal cost) on domestic two-part tariffs in 1937-38, in pence per KWH'}
    \item{gas6}{The marginal price of gas in 1935-36, in pence per therm}
    \item{gas8}{The marginal price of gas in 1937-38, in pence per therm}
    \item{kwh}{Consumption on domestic two-part tariffs per consumer in 1937-38, in kilowatt hours}
    \item{cap}{The average holdings (capacity) of heavy electric equipment bought on hire purchase (leased) by domestic two-part consumers  in 1937-38, in kilowatts}
    \item{expen}{The average total expenditure on electricity by two-part consumers  in 1937-38, in pounds}
  }
}
\source{Original article}
\references{
Houthakker, H. S. (1951) \dQuote{Some Calculations on Electricity Consumption in Great Britain}, \emph{Journal of the Royal Statistical Society. Series A (general)}, \bold{114(3)}, 359--371.

Berndt, Ernst R.(1991) The practise of econometrics, classic and contemporary, Addison-Wesley.

}
\keyword{datasets}
