\name{urban_gradient}
\docType{data}
\alias{urban_gradient}
\title{Density and Distance to the Cbd in Alabama}
\description{a cross-section of 2315 blocks from 1999

    \emph{country} :  Alabama	 

    \emph{package} : 

}
\usage{data(urban_gradient)}
\format{A dataframe containing:
  \describe{
    \item{msa}{msa name}
    \item{county}{county name}
    \item{tract}{tract id}
    \item{blkg}{blkg id}
    \item{area}{area in square km}
    \item{population}{number of inhabitants}
    \item{distance}{distance to the CBD in km}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Duranton, Gilles and Diego Puga (2020) \dQuote{The Economics of Urban Density}, \emph{Journal of Economic Perspectives}, \bold{34(3)}, 3--26.

}
\keyword{datasets}
