\name{us_counties}
\docType{data}
\alias{us_counties}
\title{Us Counties}
\description{a cross-section of 3141 counties

    \emph{country} :  United States 

    \emph{package} : 

    \emph{JEL codes: }

}
\usage{data(us_counties)}
\format{A dataframe containing:
  \describe{
    \item{fips}{fips code}
    \item{gid}{gadm code}
    \item{state}{state name}
    \item{county}{county name}
    \item{geometry}{a sfc containing the coordinates of counties border}
  }
}
\source{gadm of the geodata package}
\references{
}
\keyword{datasets}
