\name{watermelon}
\docType{data}
\alias{watermelon}
\title{Equilibrium on the Watermelon Market}
\description{a cross-section of 22 market from 1930 to 1951

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(watermelon)}
\format{A dataframe containing:
  \describe{
    \item{year}{the year, from 1930 to 1951}
    \item{q}{crop of watermelons available for harvest (millions) in log}
    \item{h}{watermelon actually harvested (millions) in log}
    \item{y}{income in log}
    \item{n}{population in log}
    \item{p}{average farm price of watermelon (dollars per thousand) in log}
    \item{pc}{average annual net farm price per pound of cottons (dollars) in log}
    \item{pv}{average farm price of vegetables (index) in log}
    \item{w}{wage in log}
    \item{pf}{transportation cost in log}
    \item{d1}{dummy for government cotton acreage allotment program, 1 for the 1934-1951 period}
    \item{d2}{dummy for world war 2, one for 1943-1946}
  }
}
\source{Stewart (2019) page 120}
\references{
Suits, D.B. (1955) \dQuote{An Econometric Model of the Watermelon Market}, \emph{Journal of Farm Economics}, \bold{NA}, 237-251.

Wold, H.O.A. (1958) \dQuote{A Case Study of Interdependent Versus Causal Chain Systems}, \emph{Review of the International Statistical Institute}, \bold{NA}, 5-25.

Stewart, Kenneth G. (2019) \dQuote{Suit's Watermelon Model: the Missing Simultaneous Equations Empirical Application}, \emph{Journal of Economics Teaching}, \bold{NA}, 115-138.

}
\keyword{datasets}
